server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
supabase>=2.0.0
redis>=5.0.0
asyncpg>=0.29.0
httpx>=0.25.0
python-dotenv>=1.0.0
